Allow tall images to push down tip text and content
In a questionWrapper, we now use a grid layout and place the original
content, the transparent overlay, and the optional image all in the same
top cell. This has the effect of autosizing the top component to the
tallest child, which pushes down the tip text and the rest of the
page content.

[#166303419]