Use opacity to set stacking order of image overlay
We have a grid layout with up to four elements placed in the same cell:

1. iFrame
2. Overlay transparency
3. Image overlay
4. Sticky note

In the naive case, the iFrame is a heavyweight object, and so will be
stacked on top of all the other components:

1. Overlay transparency
2. Image overlay
3. Sticky note
4. iFrame

However, there are two ways to create a new stacking context: positioned
elements and elements with opacity < 1 (see
https://philipwalton.com/articles/what-no-one-told-you-about-z-index/).
Since the overlay is semi-transparent and the sticky note is absolutely
positioned, we actually get

1. Image overlay
2. iFrame
3. Overlay transparency
4. Sticky note

This is close, but the image is at the back. The solution is to set the
images opacity to 0.99, placing it in the same sorting stack as the
overlay transparency:

1. iFrame
2. Overlay transparency
3. Image overlay
4. Sticky note

By this we can also see that the z-index set on the transparent overlay
was unnecessary, so it's been removed.

Bug fixed noticed in [#164012051]