Fix issues with mediaCenter layout in TE window shades
PT Story: https://www.pivotaltracker.com/story/show/182597761

[#182597761]

Currently, media in a TE window shade that uses the `mediaCenter` layout option can extend beyond the boundaries of its container in both runtime and authoring. It can make the authoring form unusable because the preview in the authoring form can take over the height of the authoring modal. These changes ensure that the media won't ever be larger than its container, and scales the media in the context of authoring. The latter is a temporary fix until we move the preview into a separate column in the authoring form.