Add vertical scrollbar to TE edit form
PT Story: https://www.pivotaltracker.com/story/show/183407024

The required scrollbar was already present. The issue was the preview section of the form was fixed in place. So when it contains a lot of content, the form fields would scroll behind the preview and weren't visible.

These changes make it so the preview will scroll out of view when the author scrolls down the edit form. This is a temporary solution until we can devote more time to putting the preview in a separate column. See https://www.pivotaltracker.com/story/show/182732380